refactor: Update TopGG constructor and perform statistics posting
This commit refactors the TopGG constructor to include clientId, apiToken, and webhookToken parameters, and modifies the postStatistics function to only use the ValidationUtil from utilities. The hasVoted function has also been refactored to use the performRequest function by passing in a query object. Additionally, the Request class now includes payload data when throwing a RequestError object. Lastly, the Webhook class now sends a successful response when receiving a vote event and the id of the bot list being used for the request is included in the response data.